+++ src/Phan/Language/EmptyUnionType.php
@@ -1058,6 +1058,11 @@ final class EmptyUnionType extends UnionType
         return false;
     }
 
+    public function hasStringType() : bool
+    {
+        return false;
+    }
+
     public function hasNonNullStringType() : bool
     {
         return false;