+++ src/Phan/Analysis/BinaryOperatorFlagVisitor.php
@@ -7,14 +7,18 @@ use Phan\AST\Visitor\FlagVisitorImplementation;
 use Phan\CodeBase;
 use Phan\Language\Context;
 use Phan\Language\UnionType;
+use Phan\Language\Type;
 use Phan\Language\Type\ArrayType;
 use Phan\Language\Type\BoolType;
 use Phan\Language\Type\FloatType;
 use Phan\Language\Type\LiteralStringType;
 use Phan\Language\Type\IntType;
+use Phan\Language\Type\MixedType;
+use Phan\Language\Type\ScalarType;
 use Phan\Language\Type\StringType;
 use Phan\Issue;
 use ast\Node;
+use Closure;
 
 // TODO: Improve analysis of bitwise operations, warn if non-int is provided and consistently return int if it's guaranteed
 class BinaryOperatorFlagVisitor extends FlagVisitorImplementation
@@ -77,8 +81,8 @@ class BinaryOperatorFlagVisitor extends FlagVisitorImplementation
         );
         static $int_or_float = null;
 
-        if ($left->isType(ArrayType::instance(false))
-            || $right->isType(ArrayType::instance(false))
+        if ($left->isExclusivelyArray()
+            || $right->isExclusivelyArray()
         ) {
             Issue::maybeEmit(
                 $this->code_base,
@@ -93,7 +97,15 @@ class BinaryOperatorFlagVisitor extends FlagVisitorImplementation
         } elseif ($left->hasType(FloatType::instance(false))
             || $right->hasType(FloatType::instance(false))
         ) {
-            if ($left->hasNonNullIntType() && $right->hasNonNullIntType()) {
+            if ($left->hasTypeMatchingCallback(
+                function (Type $type) : bool {
+                    return !($type instanceof FloatType);
+                }
+            ) && $right->hasTypeMatchingCallback(
+                function (Type $type) : bool {
+                    return !($type instanceof FloatType);
+                }
+            )) {
                 return $int_or_float ?? ($int_or_float = new UnionType([
                     IntType::instance(false),
                     FloatType::instance(false)
@@ -425,6 +437,43 @@ class BinaryOperatorFlagVisitor extends FlagVisitorImplementation
     }
 
     /**
+     * @param Node $node with type AST_BINARY_OP
+     * @param Closure(Type):bool $is_valid_type
+     * @return void
+     */
+    private function warnAboutInvalidUnionType(
+        Node $node,
+        Closure $is_valid_type,
+        UnionType $left,
+        UnionType $right,
+        string $left_issue_type,
+        string $right_issue_type
+    ) {
+        if (!$left->isEmpty()) {
+            if (!$left->hasTypeMatchingCallback($is_valid_type)) {
+                Issue::maybeEmit(
+                    $this->code_base,
+                    $this->context,
+                    $left_issue_type,
+                    $node->children['left']->lineno ?? $node->lineno,
+                    $left
+                );
+            }
+        }
+        if (!$right->isEmpty()) {
+            if (!$right->hasTypeMatchingCallback($is_valid_type)) {
+                Issue::maybeEmit(
+                    $this->code_base,
+                    $this->context,
+                    $right_issue_type,
+                    $node->children['right']->lineno ?? $node->lineno,
+                    $right
+                );
+            }
+        }
+    }
+
+    /**
      * @param Node $node
      * A node to check types on
      *
@@ -433,29 +482,19 @@ class BinaryOperatorFlagVisitor extends FlagVisitorImplementation
      */
     public function visitBinaryAdd(Node $node) : UnionType
     {
+        $code_base = $this->code_base;
+        $context = $this->context;
         $left = UnionTypeVisitor::unionTypeFromNode(
-            $this->code_base,
-            $this->context,
+            $code_base,
+            $context,
             $node->children['left']
         );
 
         $right = UnionTypeVisitor::unionTypeFromNode(
-            $this->code_base,
-            $this->context,
+            $code_base,
+            $context,
             $node->children['right']
         );
-
-        static $float_type = null;
-        static $array_type = null;
-        static $int_or_float_union_type = null;
-        if ($int_or_float_union_type === null) {
-            $float_type = FloatType::instance(false);
-            $array_type = ArrayType::instance(false);
-            $int_or_float_union_type = new UnionType([
-                IntType::instance(false),
-                $float_type
-            ]);
-        }
 
         // fast-track common cases
         if ($left->isNonNullIntType() && $right->isNonNullIntType()) {
@@ -469,6 +508,30 @@ class BinaryOperatorFlagVisitor extends FlagVisitorImplementation
                 return $left;
             }
             return ArrayType::combineArrayTypesOverriding($left, $right);
+        }
+
+        $this->warnAboutInvalidUnionType(
+            $node,
+            static function (Type $type) : bool {
+                // TODO: Stricten this to warn about strings based on user config.
+                return $type instanceof ScalarType || $type instanceof ArrayType || $type instanceof MixedType;
+            },
+            $left,
+            $right,
+            Issue::TypeInvalidLeftOperandOfAdd,
+            Issue::TypeInvalidRightOperandOfAdd
+        );
+
+        static $float_type = null;
+        static $array_type = null;
+        static $int_or_float_union_type = null;
+        if ($int_or_float_union_type === null) {
+            $float_type = FloatType::instance(false);
+            $array_type = ArrayType::instance(false);
+            $int_or_float_union_type = new UnionType([
+                IntType::instance(false),
+                $float_type
+            ]);
         }
 
         if ($left->isNonNullNumberType() && $right->isNonNullNumberType()) {
@@ -501,8 +564,8 @@ class BinaryOperatorFlagVisitor extends FlagVisitorImplementation
                 )
             ) {
                 Issue::maybeEmit(
-                    $this->code_base,
-                    $this->context,
+                    $code_base,
+                    $context,
                     Issue::TypeInvalidRightOperand,
                     $node->lineno ?? 0
                 );
@@ -511,8 +574,8 @@ class BinaryOperatorFlagVisitor extends FlagVisitorImplementation
                 && !$left->canCastToUnionType($array_type->asUnionType())
             ) {
                 Issue::maybeEmit(
-                    $this->code_base,
-                    $this->context,
+                    $code_base,
+                    $context,
                     Issue::TypeInvalidLeftOperand,
                     $node->lineno ?? 0
                 );
@@ -525,6 +588,114 @@ class BinaryOperatorFlagVisitor extends FlagVisitorImplementation
         }
 
         return $int_or_float_union_type;
+    }
+
+    /**
+     * Analyzes the result of a floating-point or integer arithmetic operation.
+     * The result will be a combination of 'int' or 'float'
+     *
+     * @param Node $node
+     * A node to check types on
+     *
+     * @return UnionType
+     * The resulting type(s) of the binary operation
+     */
+    private function getTypeOfNumericArithmeticOp(Node $node) : UnionType
+    {
+        $code_base = $this->code_base;
+        $context = $this->context;
+        $left = UnionTypeVisitor::unionTypeFromNode(
+            $code_base,
+            $context,
+            $node->children['left']
+        );
+
+        $right = UnionTypeVisitor::unionTypeFromNode(
+            $code_base,
+            $context,
+            $node->children['right']
+        );
+
+        // fast-track common cases
+        if ($left->isNonNullIntType() && $right->isNonNullIntType()) {
+            return IntType::instance(false)->asUnionType();
+        }
+
+        $this->warnAboutInvalidUnionType(
+            $node,
+            static function (Type $type) : bool {
+                // TODO: Stricten this to warn about strings based on user config.
+                return $type instanceof ScalarType || $type instanceof MixedType;
+            },
+            $left,
+            $right,
+            Issue::TypeInvalidLeftOperandOfNumericOp,
+            Issue::TypeInvalidRightOperandOfNumericOp
+        );
+
+        static $float_type = null;
+        static $int_or_float_union_type = null;
+        if ($int_or_float_union_type === null) {
+            $float_type = FloatType::instance(false);
+            $int_or_float_union_type = new UnionType([
+                IntType::instance(false),
+                $float_type
+            ]);
+        }
+
+        if ($left->isNonNullNumberType() && $right->isNonNullNumberType()) {
+            if (!$left->hasNonNullIntType() || !$right->hasNonNullIntType()) {
+                // Heuristic: If one or more of the sides is a float, the result is always a float.
+                return $float_type->asUnionType();
+            }
+            return $int_or_float_union_type;
+        }
+
+        // TODO: warn about subtracting to/from non-number
+
+        return $int_or_float_union_type;
+    }
+
+    /**
+     * @param Node $node
+     * A node to check types on
+     *
+     * @return UnionType
+     * The resulting type(s) of the binary operation
+     */
+    public function visitBinarySub(Node $node) : UnionType
+    {
+        return $this->getTypeOfNumericArithmeticOp($node);
+    }
+
+    /**
+     * @param Node $node
+     * A node to check types on
+     *
+     * @return UnionType
+     * The resulting type(s) of the binary operation
+     */
+    public function visitBinaryMul(Node $node) : UnionType
+    {
+        return $this->getTypeOfNumericArithmeticOp($node);
+    }
+
+    /**
+     * @param Node $node
+     * A node to check types on
+     *
+     * @return UnionType
+     * The resulting type(s) of the binary operation
+     */
+    public function visitBinaryPow(Node $node) : UnionType
+    {
+        return $this->getTypeOfNumericArithmeticOp($node);
+    }
+
+    public function visitBinaryMod(Node $unused_node)
+    {
+        // TODO: Warn about invalid left or right side
+        return IntType::instance(false)->asUnionType();
     }
 
     /**