+++ src/Phan/Issue.php
@@ -51,6 +51,7 @@ class Issue
     const UndeclaredTypeThrowsType  = 'PhanUndeclaredTypeThrowsType';
     const UndeclaredVariable        = 'PhanUndeclaredVariable';
     const UndeclaredVariableDim     = 'PhanUndeclaredVariableDim';
+    const UndeclaredVariableAssignOp = 'PhanUndeclaredVariableAssignOp';
     const UndeclaredClassInCallable = 'PhanUndeclaredClassInCallable';
     const UndeclaredStaticMethodInCallable = 'PhanUndeclaredStaticMethodInCallable';
     const UndeclaredFunctionInCallable = 'PhanUndeclaredFunctionInCallable';
@@ -73,6 +74,10 @@ class Issue
     const TypeInvalidClosureScope   = 'PhanTypeInvalidClosureScope';
     const TypeInvalidLeftOperand    = 'PhanTypeInvalidLeftOperand';
     const TypeInvalidRightOperand   = 'PhanTypeInvalidRightOperand';
+    const TypeInvalidLeftOperandOfAdd  = 'PhanTypeInvalidLeftOperandOfAdd';
+    const TypeInvalidRightOperandOfAdd = 'PhanTypeInvalidRightOperandOfAdd';
+    const TypeInvalidLeftOperandOfNumericOp = 'PhanTypeInvalidLeftOperandOfNumericOp';
+    const TypeInvalidRightOperandOfNumericOp = 'PhanTypeInvalidRightOperandOfNumericOp';
     const TypeInvalidInstanceof     = 'PhanTypeInvalidInstanceof';
     const TypeInvalidDimOffset      = 'PhanTypeInvalidDimOffset';
     const TypeInvalidDimOffsetArrayDestructuring = 'PhanTypeInvalidDimOffsetArrayDestructuring';
@@ -299,17 +304,19 @@ class Issue
     const AccessOverridesFinalMethodPHPDoc       = 'PhanAccessOverridesFinalMethodPHPDoc';
 
     // Issue::CATEGORY_COMPATIBLE
-    const CompatibleExpressionPHP7          = 'PhanCompatibleExpressionPHP7';
-    const CompatiblePHP7                    = 'PhanCompatiblePHP7';
-    const CompatibleNullableTypePHP70       = 'PhanCompatibleNullableTypePHP70';
-    const CompatibleShortArrayAssignPHP70   = 'PhanCompatibleShortArrayAssignPHP70';
-    const CompatibleKeyedArrayAssignPHP70   = 'PhanCompatibleKeyedArrayAssignPHP70';
-    const CompatibleVoidTypePHP70           = 'PhanCompatibleVoidTypePHP70';
-    const CompatibleIterableTypePHP70       = 'PhanCompatibleIterableTypePHP70';
-    const CompatibleObjectTypePHP71         = 'PhanCompatibleNullableTypePHP71';
-    const CompatibleUseVoidPHP70            = 'PhanCompatibleUseVoidPHP70';
-    const CompatibleUseIterablePHP71        = 'PhanCompatibleUseIterablePHP71';
-    const CompatibleUseObjectPHP71          = 'PhanCompatibleUseObjectPHP71';
+    const CompatibleExpressionPHP7           = 'PhanCompatibleExpressionPHP7';
+    const CompatiblePHP7                     = 'PhanCompatiblePHP7';
+    const CompatibleNullableTypePHP70        = 'PhanCompatibleNullableTypePHP70';
+    const CompatibleShortArrayAssignPHP70    = 'PhanCompatibleShortArrayAssignPHP70';
+    const CompatibleKeyedArrayAssignPHP70    = 'PhanCompatibleKeyedArrayAssignPHP70';
+    const CompatibleVoidTypePHP70            = 'PhanCompatibleVoidTypePHP70';
+    const CompatibleIterableTypePHP70        = 'PhanCompatibleIterableTypePHP70';
+    const CompatibleObjectTypePHP71          = 'PhanCompatibleNullableTypePHP71';
+    const CompatibleUseVoidPHP70             = 'PhanCompatibleUseVoidPHP70';
+    const CompatibleUseIterablePHP71         = 'PhanCompatibleUseIterablePHP71';
+    const CompatibleUseObjectPHP71           = 'PhanCompatibleUseObjectPHP71';
+    const CompatibleMultiExceptionCatchPHP70 = 'PhanCompatibleMultiExceptionCatchPHP70';
+    const CompatibleNegativeStringOffset     = 'PhanCompatibleNegativeStringOffset';
 
     // Issue::CATEGORY_GENERIC
     const TemplateTypeConstant       = 'PhanTemplateTypeConstant';
@@ -740,6 +747,14 @@ class Issue
                 11027
             ),
             new Issue(
+                self::UndeclaredVariableAssignOp,
+                self::CATEGORY_UNDEFINED,
+                self::SEVERITY_LOW,
+                "Variable \${VARIABLE} was undeclared, but it is being used as the left hand side of an assignment operation",
+                self::REMEDIATION_B,
+                11037
+            ),
+            new Issue(
                 self::UndeclaredTypeReturnType,
                 self::CATEGORY_UNDEFINED,
                 self::SEVERITY_NORMAL,
@@ -1142,6 +1157,38 @@ class Issue
                 "Invalid operator: right operand is array and left is not",
                 self::REMEDIATION_B,
                 10016
+            ),
+            new Issue(
+                self::TypeInvalidRightOperandOfAdd,
+                self::CATEGORY_TYPE,
+                self::SEVERITY_NORMAL,
+                "Invalid operator: right operand is {TYPE} (expected array or number)",
+                self::REMEDIATION_B,
+                10070
+            ),
+            new Issue(
+                self::TypeInvalidLeftOperandOfAdd,
+                self::CATEGORY_TYPE,
+                self::SEVERITY_NORMAL,
+                "Invalid operator: left operand is {TYPE} (expected array or number)",
+                self::REMEDIATION_B,
+                10071
+            ),
+            new Issue(
+                self::TypeInvalidRightOperandOfNumericOp,
+                self::CATEGORY_TYPE,
+                self::SEVERITY_NORMAL,
+                "Invalid operator: right operand is {TYPE} (expected number)",
+                self::REMEDIATION_B,
+                10072
+            ),
+            new Issue(
+                self::TypeInvalidLeftOperandOfNumericOp,
+                self::CATEGORY_TYPE,
+                self::SEVERITY_NORMAL,
+                "Invalid operator: left operand is {TYPE} (expected number)",
+                self::REMEDIATION_B,
+                10073
             ),
             new Issue(
                 self::TypeParentConstructorCalled,
@@ -2604,6 +2651,22 @@ class Issue
                 "Using '{TYPE}' as object will be a syntax error in PHP 7.2 (object becomes a native type that accepts any class instance).",
                 self::REMEDIATION_B,
                 3010
+            ),
+            new Issue(
+                self::CompatibleMultiExceptionCatchPHP70,
+                self::CATEGORY_COMPATIBLE,
+                self::SEVERITY_CRITICAL,
+                "Catching multiple exceptions is not available before PHP 7.1",
+                self::REMEDIATION_B,
+                3011
+            ),
+            new Issue(
+                self::CompatibleNegativeStringOffset,
+                self::CATEGORY_COMPATIBLE,
+                self::SEVERITY_CRITICAL,
+                "Using negative string offset is not available before PHP 7.1 (emits an 'Uninitialized string offset' notice)",
+                self::REMEDIATION_B,
+                3012
             ),
 
             // Issue::CATEGORY_GENERIC