+++ src/Phan/Phan.php
@@ -27,14 +27,14 @@ class Phan implements IgnoredFilesFilterInterface
     public static $printer;
 
     /** @var IssueCollectorInterface */
-    private static $issueCollector;
+    private static $issue_collector;
 
     /**
      * @return IssueCollectorInterface
      */
     public static function getIssueCollector() : IssueCollectorInterface
     {
-        return self::$issueCollector;
+        return self::$issue_collector;
     }
 
     /**
@@ -45,7 +45,7 @@ class Phan implements IgnoredFilesFilterInterface
     public static function setIssueCollector(
         IssueCollectorInterface $issueCollector
     ) {
-        self::$issueCollector = $issueCollector;
+        self::$issue_collector = $issueCollector;
     }
 
     /**
@@ -380,7 +380,7 @@ class Phan implements IgnoredFilesFilterInterface
             }
 
             // Get a count of the number of issues that were found
-            $issue_count = count((self::$issueCollector)->getCollectedIssues());
+            $issue_count = count((self::$issue_collector)->getCollectedIssues());
             $is_issue_found =
                 0 !== $issue_count;
 
@@ -498,7 +498,7 @@ class Phan implements IgnoredFilesFilterInterface
      */
     private static function display()
     {
-        $collector = self::$issueCollector;
+        $collector = self::$issue_collector;
 
         $printer = self::$printer;