+++ NEWS.md
@@ -22,6 +22,7 @@ New features(Analysis)
   New issue types: `PhanAccessNonStaticToStaticProperty`, `PhanAccessStaticToNonStaticProperty`
 + Warn about inheriting from a class/trait/interface that has multiple possible definitions (#773)
   New issue types: `PhanRedefinedExtendedClass`, `PhanRedefinedUsedTrait`, `PhanRedefinedInheritedInterface`
++ Infer more accurate types for the side effects of assignment operators (i.e. `+=`, `.=`, etc) and other binary operations.
 
 Maintenance:
 + Update signature map with more accurate signatures (#1761)